split export modules
